Chart of new moon observations
These are based on the Kahun Papyri as given in the same chart in a paper entitled "The Kahun Papyrus and the Date of the Twelfth Dynasty (With a Chart)" by Lynn H. Wood published in [* *Bulletin of the American Schools of Oriental Research No. 99 (Oct., 1945)* *, pp. 5-9](https://www.jstor.org/stable/1355209).